Scope the `transactions_total` to a Report & that Report's date range
The Report CSV will always contain all Activities associated with a
Fund/Organisation pair, whether they had transactions in the report period or
not. But the transaction totals must be scoped to a single Report and its date
range, not be the totals for that Activity for all time.

The new method is called `actual_total_for_report_financial_quarter` for clarity