Calculate the total of all Planned Disbursements in a Report's date range
Planned Disbusements scoped to a single Report and that Report's date
range are summed and returned in forecasted_total_for_report_financial_quarter

The Report CSV will always contain all Activities associated with a
Fund/Organisation pair, whether they had planned disbursements in the report
period or not. But the planned disbursement total must be scoped to a single
Report and its date range, not be the totals for that Activity for all time.