Update the Report policy
We already have a good understanding of the what state transitions are
valid and which user can make them.

This work updates the report policy and related spec to make it possible
to enforce this understanding and support future work to allow users to
move a report though its state inside the application.

Co-authored-by: jcoglan <james.coglan@dxw.com>