Reports can be set to active
When the application creates a new report, its state defaults to
'inactive' this gives BEIS control of when the Delivery Partners can see
and interact with the reports.

The goal is to keep the time between these two events as short as
possible but it is good to have control of when the reports become
'active'.

This works enables BEIS users to move reports from the inactive state to
the active state. Part of this meant updating the report locales to
allow for more states whilst keeping the structure.