Use memoisation on Activity#variance_for_report_financial_quarter
Adjust the tests to work more effectively with memoisation. Previously we were
using one `it` block with multiple time periods, and as a result this memoisation
method doesn't work (the memoised result for all the totals was used in all tests
in the `it` block).

So I have adjusted the tests to use one time period in each `it` block, this also
represented a more realistic use of the code (we will not be instantiating one
instance of the Activity to be used across multiple time periods and reports).