`Call dates` field added to the activity form
As part of the additional fields needed in RODA, and having identified this one
as a priority for the users, we have now added `call open date` and `call close
date` to the service. These dates are only present in level C and D activities,
and they relate to the timeline to apply for grants (they are not the same as
IATI activity dates). As requested by the client, this new field comes in two
different steps. On a first step, the DP user will select whether an activity
has any calls. If the answer to this is `yes`, then the second step will be shown,
in which the DP can select both `call open date` and `call close date`. If
there are no calls for that activity, the second step will be ignored.

This field would be mandatory for all level C and D activities created through
RODA, but optional for those `ingested: true`.

This commit includes the new views, the necessary changes on the DB, on the
controller, activity presenters, locales and the tests to assert we get the
desired behaviour from this new feature.