Modify a spec due to a failing test
After creating the new fields `call open date` and `call close date` and adding
the necessary validations we had a failing test on the `ingest_iati_activities`
spec due to a project being created without `ingested` field and with
`call_present` being nil. I decided to add `call_present: false` instead of
`ingested: true` because I think this fits better, and also to avoid adding a
`ingested` field that could result in an unexpected change on the test behaviour.