Use `t()` over `I18n.t()` in controllers and views
I am about to rename a database column, which necessitates renaming the
associated keys in the locale files. I want to make sure I've caught all
these, and it's currently easy for a test like this to pass by accident:

    expect(page).to have_content I18n.t("action.fund.create.success")

This is because I18n.t does not throw an error when given an unknown
key, unless you also pass `raise: true` to it; it returns a string like
"translation missing: action.fund.create.success". If your view is also
calling I18n.t with this key then the test will pass, even though the
page's content is broken.

To make it more likely this typoe of bug doesn't happen, I'm switching
to using the ActionView `t()` wrapper [1] in almost all places. This has
the advantage that it can be configured to raise an error on unknown
keys, meaning the above test would fail if I've renamed a key and not
updated the template appropriately. This behaviour is enabled by setting
`raise_on_missing_translations` in the `test` environment.

The development workflow for adding new copy is unchanged; you can still
render a page with missing translations in development without the
entire page failing, but when running tests an unknown translation will
cause a test failure.

There are some places I have _not_ replaced I18n.t with `t()`. There are
two reasons that any remaining call to I18n has not been replaced:

The first is that it is called from somewhere that's not a controller or
view, so `t()` is not defined in that context.

The second reason is that it is used with a dynamically generated key
where the dynamic part affects the _end+ of the key. The ActionView
`t()` wrapper has an extra behaviour that I18n.t does not have: if the
key ends with `_html` then the result is automatically marked as
`html_safe` and so will not be HTML-encoded. This introduces a risk of
XSS if the resulting string has not been checked for HTML safety, so I
would rather keep using I18n.t in those cases and have the result
HTML-encoded by default.

This conversion is not exhaustive and does not promise to catch all
translation errors; it's a best-effort to reduce the risk of missing a
rename somewhere. Note that `t()` in _controllers_ is a simple wrapper
around I18n.t while `t()` in _views_ (and therefore in helpers) has
additional behaviour as documented at the link below. So, missing
translations in controllers will not be caught, but views are a much
bigger risk in terms of renaming keys so this feels like a safety win.

[1]: https://api.rubyonrails.org/classes/ActionView/Helpers/TranslationHelper.html#method-i-translate