Add longform function to get the next 4 financial quarters for a Report
I'm hoping this can be done in a shorter form later.

A method to return the next four financial quarters as an array of strings, e.g.
"Q1 2020" based on the quarter of the current report.

E.g. if the current report is for Q1 2020, the next 4 are Q2 2020, Q3 2020,
Q4 2020 and Q1 2021