Rename Activity#identifier to `delivery_partner_identifier`
We have determined a need for the service to model three distinct
identifiers for an Activity:

- Delivery Partner ID: this is the DP's own identifier for the activity,
  and should be displayed so the DP can cross-reference with their own
  records. It can contain any characters and be of arbitrary length.

- RODA ID: this is the canonical ID within RODA and which we export to
  external destinations like DfID. As such it has additional
  restrictions on length.

- Transparency ID: this is the ID we use in exports to IATI and is
  formed by prefixing the RODA ID with `GB-GOV-13-`, and converting any
  non-compatible characters to hyphens.

We currently have a sort of hybrid of these in the system. Activities
are arranged in a hierarchy of levels, with BEIS owning levels A and B
and DPs owning C and D:

    A: fund
    B: programme
    C: project
    D: third party project

Each activity has an `identifier` field which we use as a _component_ of
the whole when generating the Transparency ID; the identifiers of an
activity and all its ancestors are concatenated:

    Transparency ID = GB-GOV-13-{A}-{B}-{C}-{D}

But on ingest of historical data, we do not (and cannot) parse IDs into
components like this, because the components parts can themselves
contain hyphens. So, we store whatever ID appears in the ingested data
as the `identifier`, with the string `GB-GOV-13-` and a possible fund
name (level A component) removed from the start. This string is a
best-effort inference at what the Delivery Partner's ID is.

So we have determined that the IDs already in the system are closer to
being Delivery Partner IDs than RODA IDs: they're arbitrary strings, not
necessarily stored as level-component parts, and might exceed length
limits required by DfID and others. Therefore we are renaming
`identifier` to `delivery_partner_identifier`, and will add a new field
to model the RODA ID afterward. After we've done that:

- The Delivery Partner ID should be editable -- it's whatever the DP
  wants it to be and has no consistency contraints within RODA

- The RODA ID should not be editable -- it needs to be stable, stored in
  level-component parts, and will be used to generate the Transparency
  ID which is exported to IATI

This commit renames the database column, and also renames the keys in
config/locales/models/activity.yml. Some of these keys are used
dynamically based on attribute names, either by the app or by Rails
itself, and some are used statically, but I thought it best to use a
consistent name for the field throughout.