Country to region mapping changes
Following changes on the region codes to match those IATI provides, the
`country_to_region_mapping.yml` is no longer accurate. This commit introduces
the necessary changes for a up-to-date mapping inline with the new country list
provided by the client.

In addition, there were a few failing tests due to changes on the mapping that
have been amended on this commit.