Fix for `Add activity` buttons
On a previously merged PR we intended to replace all `Create activity` buttons
for `Add activity` buttons, as recommended by the content review. Some of the
buttons still appeared as `Create activity` due to an unchanged field on a yml
file. This commit introduces the changes to the `create_activity` field.