Enable a BEIS user to mark a submitted report as "in review"
This is a basic first pass at moving Reports from "submitted" to "in review"

When a Delivery Partner submits a report, it appears on the BEIS user's
dashboard. The BEIS user can then mark as it being "in review" to indicate to
other BEIS users that the report is being looked at.

The UI is very basic and will be iterated upon.