Add `activities.roda_identifier_{fragment,compound}`
This begins the work to model the "RODA Identifier". Whereas we
currently have a "Delivery Partner Identifier" in the system, the RODA
Identifier will be the canonical ID within RODA and be the basis for
external reporting to DfID, IATI, etc.

The ID must be strictly hierarchical, for example, all the programmes
under the GCRF fund would have the prefix `GCRF-`. If that programme's
ID is `GCRF-Prog`, then all projects under it would have the prefix
`GCRF-Prog-`, etc.

So, we model the identifier as each activity owning a "fragment" of the
ID, and the whole "compound" ID is made up by concatenating the
fragments from each level in the hierarchy.

We will need to cache the "compound" form of the ID so that it's easy to
look up an activity by its complete RODA Identifier, for example to
implement bulk upload of transaction data. It also makes it easier to
ensure that all RODA Identifiers are unique.