Set RODA Identifier during activity creation
This adds a step to the activity creation wizard for setting a RODA
Identifier fragment. This is mostly boilerplate just to add the step,
and amend any tests that need to take the new step into account.

The only interesting piece of logic here is that when the RODA
Identifier fragment is set, we call `Activity#cache_roda_identifier!` to
fix its compound RODA Identifier in the database; the field should
disallow modification after this happens.