Skip adding a RODA Identifier if it's not possible
Adding a RODA Identifier to an activity should only be allowed if both
the following are true:

- the activity does not have `roda_identifier_fragment` set
- none of the its ancestors is missing `roda_identifier_fragment`

This forces you to work down the tree, first adding Identifiers to
funds, then to programmes, and so on. There are two main reasons to
enforce this.

First, the RODA Identifier will have validation rules about its length
that depend on the length of the parent's Identifier fragment, in the
case of level B and D activities. This means activities at those levels
cannot be meaningfully validated without their parent having a RODA
Identifier set. The alternative is to allow RODA Identifiers to be
completed in any order, but that would require us to check all the
children's identifiers are still valid after you set a level A or C
Identifier.

Second, a similar argument applies to caching. The
`roda_identifier_compound` field is a cache of the combined
`roda_identifier_fragment` values for an activity and its ancestors.
Forcing the IDs to be filled out from the root downwards, and making the
field uneditable, means we never have to invalidate that cache on an
activity and all its descendents.