Fix AMS GCRF programmes hierarchy
During the previous data migration we moved activities from projects to
programmes to reflect mapping changes.

Unfortunately, we didn't change the parent mappings, which lead to
programmes being the parent of other programmes, which shouldn't be
possible.

This data migration updates the affected activities to link them
directly to the GCRF fund activity instead.