Report shows top level financial details
Eventually the report views will show all of the changed data for a
report. As a first step toward this we want to show one of the key indicators,
the totals for forecasted spend and actual spend and the variance
between the two.

As these are already calculated for the csv file we don't have to do
much to show them in the applications user interface.