Forbid setting a compound RODA identifier if it already exists
The policy for setting this cached value should usually be dealt with
by calling `Activity#can_set_roda_identifier?` before attempting to set
`roda_identifier_fragment` and calling `cache_roda_identifier!`.

To catch situations where this field is set when it should not be, we
add a custom setting method that blocks it being set if it already has a
value.