Temporary solution for `ingest data` failing tests
After including `intended beneficiaries` as a new field required in some
situations, I got a bunch of failing tests regarding the ingest of existing
activities. After discussion with other members of the dev team, we agreed on
using a placeholder as a temporary solution until we have the real data to be
ingested and we can work out a better way to ingest it. For the optional step
of `requires_additional_benefitting_countries` we default the value to `true`.
This will allow us to ingest intended beneficiaries in the future and keeps the
tests passing.