Derive `transparency_identifier` from `roda_identifier_compound`
Previously, the Transparency Identifier was derived from the Delivery
Partner Identifier, but concatenating the identifiers from all the
activities from level A downwards. This doesn't work because Delivery
Partner Identifiers are not stored in this hierarchical way; they're
inferred from an activity's ID when we ingest historical data, or set by
the delivery partner.

We want to use the canonical RODA Identifier to set the Transparency
Identifier. Transparency Identifiers should only contain letters,
digits, and hyphens, so we convert any string of characters not in one
of these categories to hyphens when converting the RODA Identifier.

We also change where this change happens. Since it's derived from the
RODA Identifier, we do this inside `Activity#cache_roda_identifier`, and
not in the form step where the Delivery Partner Identifier is set.