Allow Delivery Partner Identifiers to be edited
We previously made this field uneditable, because it is used to generate
the Transparency Identifier. This is now done based on the RODA
Identifier, which is not editable after first being set. The Delivery
Identifier is now purely for the DP's own use and can safely be edited.

This effectively reverts 9abf61d0 (Do not allow the user to edit the
activity identifer after it is created, 2020-07-24).