Reports cannot be unique for the Level A activity
A report is for an organisation and a level A activity (Fund). We were
validating that reports were unique at level A for a given organisation
using `validates_uniqueness_of` and a scope to organisation, this stops
a new report being created for the organisation for the same level A
activity.

Multiple reports for the same organisation and level A activity must
exist so this validation is incorrect and has been removed.