Do no try to activate invalid reports
Because reports exist before they are valid it was possible for a user
to attempt to activate an invalid one, essentially a report with no
`description`.

As a short term fix to stop the application failing, we handle this case
and show an error.

For this to make some sense we force the user to activate a report from
the report show page rather than a shortcut on the report index page,
this means we can always return the user to the same place in the
application should a user find themselves in this position.