Rebuild transparency identifiers from existing data
We have just changed the way Transparency Identifiers are generated;
they're produced based on the RODA Identifier, not the Delivery Partner
Identifier. This means the Transparency Identifiers currently in the DB
will not be consistent.

Before doing this, we checked whether this would affect the identifiers
being used to publish anything to IATI. An activity will be affected if
it has no `previous_identifier` but does have a
`transparency_identifier`. All the records that are in this state (of
which there are 129 at the time of writing) are records that were _not_
originally ingested from IATI. The full data for all combinations of
these fields is:

    ingested  | has previous_identifier | has transparency_identifier | count
    ----------+-------------------------+-----------------------------+------
    no        | no                      | no                          | 4
    no        | no                      | yes                         | 129
    no        | yes                     | no                          | 0
    no        | yes                     | yes                         | 1
    yes       | no                      | no                          | 0
    yes       | no                      | yes                         | 0
    yes       | yes                     | no                          | 50
    yes       | yes                     | yes                         | 264

These 129 records have not yet been pyublished to IATI, and so changing
their `transparency_identifier` is not dangerous. Any record that has a
`transparency_identifier`, and _also_ has a `previous_identifier`, can
have its `transparency_identifier` changed or removed, because
`previous_identifier` will be used for publishing it to IATI.

Therefore, this data migration removes all the existing
`transparency_identifier` values from the database, and regenerates
those for any records that have had their RODA Identifier set, using the
same logic from the model's `cache_roda_identifier` method. The lock
placed on the records means no concurrent transaction can update the
`transparency_identifier` while this migration is running.