New reports are created automatically
When the current report is approved we go ahead and create the next report
in the inactive state, ready for BEIS to activate.

The application relies on there only ever being one active report for a
given delivery partner organisation and level A activity so we want the
application to handle the creation of the new one to avoid error.

We only ever create the new report if one does not already exist in any
state other than `approved` for the organisation and fund.

We also know that due to the nature of the service a report will have to
be approved either in the financial quarter in which it was created or
the following one - in both cases creating the new report upon approval
means that the new report has the correct financial year.