Pass report state as a parameter
To make sure that the ReportsStateController takes an action that
matches the user's intent, the new state should be passed a parameter
rather than having the controller infer the new state from the report's
current state.

The current implementation allows for a race condition where the
following sequence of events happens:

0. There is a report currently in the 'active' state.

1. Alice loads the 'confirm' page, for example to prepare to submit the
   report.

2. Bob loads the same page; both Alice and Bob will see a page asking
   them to confirm they want to submit the report.

3. Alice submits the form. The report is currently active so it gets
   marked as submitted.

4. Bob submits the form. The report on the server is now submitted, so
   this is interpreted as an attempt to mark the report as 'in review'.

Bob's request is thus mis-handled; either he is not allowed to mark the
report as 'in review' and the request fails, or he is allowed and an
unintended state change happens.

This is mitigated by adding a `state` param to the form, so that the
client is explicitly indicating what next state it's requesting. The
server then validates whether this request is actionable, rather than
guessing based on the current report state.