Make report state changes idempotent
In a previous commit we fixed a race condition that happens in the
following sequence of events:

0. There is a report currently in the 'active' state.

1. Alice loads the 'confirm' page, for example to prepare to submit the
   report.

2. Bob loads the same page; both Alice and Bob will see a page asking
   them to confirm they want to submit the report.

3. Alice submits the form. The report is currently active so it gets
   marked as submitted.

4. Bob submits the form. The report on the server is now submitted, so
   this is interpreted as an attempt to mark the report as 'in review'.

Passing `state` as a param removes the ambiguity of Bob's request in
step 4: it is now a request to mark the report as submitted. But the
report is already in the state Bob requested, following Alice's request
in step 3.

If the report is already in the state requested by the client, then no
changes need to be made and the success page be rendered. To make sure
no concurrent request can modify the report while we're validating the
state transition, we lock the record during the `update` action. This
also makes sure the user sees the right outcome if they double-submit
the confirmation form.