`ODA eligibility` form step added to activity
As part of the new fields on create activity journey requested by the client,
the ODA eligibility has been added to every level activity. Worth noting that
`ODA eligibility` is a requirement only for levels C and D activities, but the
client suggested we implemented this on all levels for now. They are happy with
refining this later if needed.

This new field has two options: `Eligibile` and `No longer eligible`. A defaulted
to `Eligible` has been implemented, as requested by the client. This field does
not get reported to IATI or included in any XML.

This commit introduces the new form step, as well as tests to assert we achieve
the desired behaviour.