Update the Activity policy
The activity policy needed to account for the state of the report.

Following the Report policy we also refactored the spec to cover
everything we think the policy should permit or deny. This led to
extensive refactoring!

The goal here was to document in the spec all states that affect an
activity's policy regardless of if we use it currently in the
application, this may lead to some parts of the UI leading to 'not
authorised' pages for the moment.

The activity policy is more complicated due to users being able to
create an activity before assigning it a level or parent. Without a
level or parent activity we cannot work our what the permissions should
be.

Because activities are so fundamental to the application, this work will
break other specs where the setup will now require a report in the
correct state, these will be fixed in the next commit to keep the size
of this one manageable.