Fix test suite
With the new activity policy in-place a lot of our specs now need a
report in the appropriate state in the setup.

Generally speaking a report always needs to be setup with a specific
organisation and fund in the setup so no change to the report factory
was made - we may want to spend more time to think about how our specs
get setup in respect to reports.