Import a single transaction from a CSV
This begins the work of importing transactions in bulk from a CSV input.
Most of the machinery will like in the ImportTransactions service, which
takes an array of hashes rather than a CSV file directly, to make it
easier to test.

The initial implementation maps all the bulk import columns to their
places in the database, but does not do any additional validation or
conversion beyond that performed by the model. It does provide some
defaults for mandatory fields:

- The currency is GBP
- The transaction type is 'Disbursement'
- The 'providing organisation' information is inherited from the
  Activity