Total applications and awards form step added
`Total applications` and `Total awards` are combined in a new field when users
create level C or D activities. They will exist only if there is a call present
on the activity. Applications refer to organisations applying for a grant/fund
related to a specific activity. Awards is related to the number of organisations
that have been awarded the grant/fund they applied for.

This field is optional for activities when `ingested: true`. It does not get
reported to IATI.

Both fields (applications and awards) only accept integers. We agreed with the
client that the value can be set to `0` whenever the application process is
still ongoing, or when an activity has been suspended before the call.