Validate the activity as part of a bulk transaction upload
This is the first piece of validation that ImportTransactions does, so
this also sets up a bunch of machinery to support error reporting.
Errors report the row offset, column name, cell value, and error
message, so errors can be easily located in a spreadsheet.

Because this class doesn't assume CSV input per se but takes an array of
hashes, the row offset is zero-based. It can be adjusted if the caller
knows more about the input format.

This first validation checks that the `Activity RODA Identifier` field
refers to an existing activity, otherwise an error is produced and the
transaction is not imported.