Validate dates on bulk-uploaded transactions
Here we introduce support for reporting errors that result from
conversions from the data format supported by the bulk upload CSV
service. The conversion methods can raise exceptions, which will be
caught and stored on the ImportTransactions instance.

The first such validation is that Date must parse as ISO-8601, according
to the Ruby implementation of this format.

The value reported in the error is the original value from the imported
row, not the value after any conversion is applied.