Refactor error handling in ImportTransactions
The `import_row` method is getting fairly complex, so here we break it
into smaller pieces, in particular not mixing error handling in with the
rest of the logic quite so much.

For each row, errors are collected in a hash, which can be updated
incrementally or import all the errors from another object, as in

    errors.update(converter.errors)

At the end of `import_row` we convert the hash's contents into `Error`
objects with row numbers attached.