Validate dates at the model layer in ImportTransactions
If the `Date` field of an imported row is empty, we return `nil` from
the conversion function rather than an empty string. This triggers a
more meaningful validation in the model.

Our imported should only validate the _formatting_ of the date, not its
content -- any semantic validation should be done by the model, this
class is merely a parser/translator.

To effect this, the result of `CreateTransaction` is extracted and we
copy any ActiveRecord validation errors into our own error set, using
the attribute name to fetch the raw value from the input.