Refactor CreatePlannedDisbursement#report and CreateTransaction#report
These two classes both infer the currently active report that a planned
disbursement or transaction should be added to. During bulk transaction
upload, this report will be provided explicitly, because we'll launch
the upload process from the page for a specific report.

When ImportTransactions is calling CreateTransaction, I would rather
pass the Report in, rather than have CreateTransaction infer the current
Report and possibly disagree with the caller. Therefore I've factored
the logic shared by these two classes into
`Report#editable_for_activity`, and made both classes call this if
they're not given a Report explicitly.

The one difference introduced here is that the previous implementation
looked only for `active` reports. We actually want to find any report
the delivery partner can currently edit, which means it may be `active`
or `awaiting_changes`.