Authorise the user for each bulk upload transaction
In normal controller actions we only need to authorise the user for a
single resource/action. In bulk upload, the uploaded CSV might name any
activity, so we need to make sure the current user is allowed to add
data for that activity, and that the activity is under the scope of the
current report.

To check the uploader is authorised to create this transaction, I'm
checking ActivityPolicy#create?, which is the same thing that
Staff::TransactionsController#create checks.

To check the activity is part of the current report, I'm checking that
their organisation and fund match, which is what would cause
`Report.editable_for_activity` to link the two objects.