Refactor `import_row` to a RowImporter object
We had built up a lot of logic in the methods of ImportTransactions that
were passing the same objects between each other. This suggested pulling
the whole process of importing a row into a distinct object, so that the
various objects involved can live in its instance variables and not have
to be passed around explicitly.

This looks like a large change but it's not meant to change any logic.
It just moves a set of logic into a new class while replacing function
parameters with instance variables.