Show the deadline on the report edit form
When the financial year and quarter were added to the edit form as
read-only values we had to use a presenter rather than the report
itself.

Something about doing so stops the value of the deadline field showing
in the form. If we switch to the report object itself,everything works
as expected so we instead pass in the report presenter as a local to the
form partial and use that to render the financial quarter and year.