Download a CSV template to begin bulk transaction upload
To initiate the bulk upload process, a user downloads a CSV that
contains the column headers we require in ImportTransactions. It lists
out the RODA Identifier for each activity that should appear in the
current report, meaning all the level C and D activities for the
report's organisation and fund, that have a RODA Identifier.

(The Delivery Partner Identifier and Name are also included in the
download for informational purposes and are ignored on upload.)

The tests include a third project that belongs to the same organisation,
but to a different fund from the other two, to make sure it's not
included in the download for this report.