Activities now display hierarchy levels on the UI
As a request from the client, in order to ensure consistency in how people view
the service, we have updated the terminology for activity levels to now show
both the former terminology (`Fund`, `Programme`, `Project`, and `Third-party
project`) along with their respective levels (A, B, C, and D).

This is only a change on the UI, mainly in the translation files, so we can
still keep our code without causing much disruption at this moment.

This commit includes changes on the translation files, amendments on the specs,
and a custom method for capitalisation when presenting the activity levels in
some of the views.