Updated definitions for activity levels
As part of the changes requested by the client regarding the activity hierarchy
terminology, the definitions for each level of activity needs to be modified too.

This commit introduces the updated definitions that will be displayed when users
click on `Add activity` and choose what level of activity they want to create.