Update transaction policy
BEIS users can add and edit transactions at any time to Fund and
Programme activities. They can never add or edit transactions on any
other activity.

Delivery partner users can add transactions to an activity as long as
there is a report in an editable state (active or awaiting changes) for
the same organisation and fund as the activity.

Delivery partner users can edit a transaction on an activity as long as
there is a report in and editable state and the report is the one in
which the transaction was added.

The goal of the spec is to document this behaviour as best it can and
test it, this leads to a lot of nested context to cover all the states.