Define editable states in Report
We are starting to see lots of places where we have the concept of a
Report in an 'editable state' - this means a report in either the active
or awaiting_changes state.

It feels safer to define this in the Report model rather than sprinkled
around the code where it could be missed if it is ever changed or
updated.