Update planned disbursement policy
Right now the planned disbursement is the same as the transaction
policy:

BEIS users can add and edit planned disbursements as long as the
activity is a programme or fund.

Delivery partners can add a new planned disbursement as long as there is
an editable report for the organisation and fund of the activity to
create the planned disbursement for.

Delivery partner user can only edit a planned disbursement when the
editable report is the one in which it was created.